Simplify presence example
At the moment, the presence example is a little bit complicated,
because it handles adding multiple clients to the same window. This also
breaks its behaviour, because presence IDs are based on an array index,
which isn't unique when opening a second window.

This change just has a single client in the window, which has the
advantage of simplifying the code.

In order to test multiple users, simply open multiple windows.